moved dropzone up one level
- moved react-dropzone component up one level
- removed currying since unncessary now
- renamed CertificateDropZone to CertificateVerificationStatus to better reflect its responsibility
