+++ src/components/CertificateDropZone/CertificateDropZoneContainer.js
@@ -1,6 +1,7 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import Router from "next/router";
+import Dropzone from "react-dropzone";
 import { connect } from "react-redux";
 import {
   updateCertificate,
@@ -15,8 +16,29 @@ import {
   getStoreStatus
 } from "../../reducers/certificate";
 import { updateNetworkId } from "../../reducers/application";
-import CertificateDropZone from "./CertificateDropZone";
+import CertificateVerificationStatus from "./CertificateVerificationStatus";
 
+const onFileDrop = (
+  acceptedFiles,
+  handleCertificateChange,
+  handleFileError
+) => {
+  // eslint-disable-next-line no-undef
+  const reader = new FileReader();
+  if (reader.error) {
+    handleFileError(reader.error);
+  }
+  reader.onload = () => {
+    try {
+      const json = JSON.parse(reader.result);
+      handleCertificateChange(json);
+    } catch (e) {
+      handleFileError(e);
+    }
+  };
+  if (acceptedFiles && acceptedFiles.length && acceptedFiles.length > 0)
+    acceptedFiles.map(f => reader.readAsText(f));
+};
 class CertificateDropZoneContainer extends Component {
   constructor(props) {
     super(props);
@@ -48,20 +70,35 @@ class CertificateDropZoneContainer extends Component {
 
   render() {
     return (
-      <CertificateDropZone
-        document={this.props.document}
-        fileError={this.state.fileError}
-        handleCertificateChange={this.handleCertificateChange}
-        handleFileError={this.handleFileError}
-        verifying={this.props.verifying}
-        issuerIdentityStatus={this.props.issuerIdentityStatus}
-        hashStatus={this.props.hashStatus}
-        issuedStatus={this.props.issuedStatus}
-        notRevokedStatus={this.props.notRevokedStatus}
-        verificationStatus={this.props.verificationStatus}
-        resetData={this.resetData.bind(this)}
-        storeStatus={this.props.storeStatus}
-      />
+      <Dropzone
+        id="certificate-dropzone"
+        onDrop={acceptedFiles =>
+          onFileDrop(
+            acceptedFiles,
+            this.handleCertificateChange,
+            this.handleFileError
+          )
+        }
+        className="h-100"
+      >
+        {props => (
+          <CertificateVerificationStatus
+            document={this.props.document}
+            fileError={this.state.fileError}
+            handleCertificateChange={this.handleCertificateChange}
+            handleFileError={this.handleFileError}
+            verifying={this.props.verifying}
+            issuerIdentityStatus={this.props.issuerIdentityStatus}
+            hashStatus={this.props.hashStatus}
+            issuedStatus={this.props.issuedStatus}
+            notRevokedStatus={this.props.notRevokedStatus}
+            verificationStatus={this.props.verificationStatus}
+            resetData={this.resetData.bind(this)}
+            storeStatus={this.props.storeStatus}
+            hover={props.isDragAccept}
+          />
+        )}
+      </Dropzone>
     );
   }
 }
